+++ packages/config/src/options.ts
@@ -681,11 +681,6 @@ export const breakingRootOptions: Array<BreakingOption> = [
     errorKey: 'CONFIG_FILE_INVALID_ROOT_CONFIG',
     isWarning: false,
     testingTypes: ['e2e'],
-  }, {
-    name: 'experimentalRunAllSpecs',
-    errorKey: 'EXPERIMENTAL_RUN_ALL_SPECS_E2E_ONLY',
-    isWarning: false,
-    testingTypes: ['e2e'],
   },
   {
     name: 'experimentalOriginDependencies',
@@ -739,11 +734,6 @@ export const testingTypeBreakingOptions: { e2e: Array<BreakingOption>, component
     {
       name: 'testIsolation',
       errorKey: 'CONFIG_FILE_INVALID_TESTING_TYPE_CONFIG_COMPONENT',
-      isWarning: false,
-    },
-    {
-      name: 'experimentalRunAllSpecs',
-      errorKey: 'EXPERIMENTAL_RUN_ALL_SPECS_E2E_ONLY',
       isWarning: false,
     },
     {