+++ system-tests/__snapshots__/protocol_spec.js
@@ -6799,7 +6799,7 @@ exports['component events - experimentalSingleTabRunMode: true'] = `
             "line": 4,
             "column": 0,
             "whitespace": "    ",
-            "stack": "Error\\n    at ./src/components/HelloEarth.cy.jsx (http://localhost:2121/__cypress/src/spec-0.js:16:1)\\n    at __webpack_require__ (http://localhost:2121/__cypress/src/main.js:108:41)"
+            "stack": "Error\\n    at ./src/components/HelloEarth.cy.jsx (http://localhost:2121/__cypress/src/spec-0.js:16:1)\\n    at __webpack_require__ (http://localhost:2121/__cypress/src/main.js:111:41)"
           },
           "retries": -1,
           "_slow": 250,
@@ -6902,7 +6902,7 @@ exports['component events - experimentalSingleTabRunMode: true'] = `
             "line": 4,
             "column": 0,
             "whitespace": "    ",
-            "stack": "Error\\n    at ./src/components/HelloMars.cy.jsx (http://localhost:2121/__cypress/src/spec-0.js:16:1)\\n    at __webpack_require__ (http://localhost:2121/__cypress/src/main.js:108:41)"
+            "stack": "Error\\n    at ./src/components/HelloMars.cy.jsx (http://localhost:2121/__cypress/src/spec-0.js:16:1)\\n    at __webpack_require__ (http://localhost:2121/__cypress/src/main.js:111:41)"
           },
           "retries": -1,
           "_slow": 250,
@@ -8658,7 +8658,7 @@ exports['component events - experimentalSingleTabRunMode: false'] = `
             "line": 4,
             "column": 0,
             "whitespace": "    ",
-            "stack": "Error\\n    at ./src/components/HelloEarth.cy.jsx (http://localhost:2121/__cypress/src/spec-0.js:16:1)\\n    at __webpack_require__ (http://localhost:2121/__cypress/src/main.js:108:41)"
+            "stack": "Error\\n    at ./src/components/HelloEarth.cy.jsx (http://localhost:2121/__cypress/src/spec-0.js:16:1)\\n    at __webpack_require__ (http://localhost:2121/__cypress/src/main.js:111:41)"
           },
           "retries": -1,
           "_slow": 250,
@@ -8761,7 +8761,7 @@ exports['component events - experimentalSingleTabRunMode: false'] = `
             "line": 4,
             "column": 0,
             "whitespace": "    ",
-            "stack": "Error\\n    at ./src/components/HelloMars.cy.jsx (http://localhost:2121/__cypress/src/spec-0.js:16:1)\\n    at __webpack_require__ (http://localhost:2121/__cypress/src/main.js:108:41)"
+            "stack": "Error\\n    at ./src/components/HelloMars.cy.jsx (http://localhost:2121/__cypress/src/spec-0.js:16:1)\\n    at __webpack_require__ (http://localhost:2121/__cypress/src/main.js:111:41)"
           },
           "retries": -1,
           "_slow": 250,