+++ system-tests/projects/run-all-specs-ct-vite/component/folder-a/component-b.cy.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import ComponentB from './component-b'
+
+describe('<ComponentB />', () => {
+  it('renders', () => {
+    cy.mount(<ComponentB />)
+    cy.get('h1').should('have.text', 'I am Component B')
+  })
+})