+++ system-tests/projects/run-all-specs-ct-vite/cypress/support/component.ts
@@ -1,0 +1,11 @@
+import { mount } from 'cypress/react'
+
+declare global {
+  namespace Cypress {
+    interface Chainable {
+      mount: typeof mount
+    }
+  }
+}
+
+Cypress.Commands.add('mount', mount)